Immigrate reference operations to a separate module
